+++ install/install-ubuntu.sh
@@ -17,6 +17,11 @@ apt-get -y upgrade
 # Install server software
 apt-get -y install openjdk-8-jdk vim
 
+# Install fonts
+# (Installing all fonts is overkill but will make sure that the necessary fonts
+# get installed)
+apt-get -y install fonts-*
+
 # Install cups 
 apt-get -y install cups hplip cups-bsd
 systemctl stop cups
@@ -26,6 +31,8 @@ chmod 644 /etc/cups/cupsd.conf
 chown root:lp /etc/cups/cupsd.conf
 systemctl enable cups
 systemctl start cups
+systemctl enable cups-browsed
+systemctl start cups-browsed
 
 # Install mariadb
 apt-get -y install mariadb-server
@@ -35,7 +42,7 @@ echo "Creating databases"
 mysql < script/database.sql
 
 # Install kumoreg
-useradd kumoreg -s /sbin/nologin
+useradd -m kumoreg -s /sbin/nologin
 usermod -a -G lpadmin kumoreg
 mkdir /usr/local/kumoreg/
 mv KumoReg-*.jar /usr/local/kumoreg/