+++ src/main/java/org/kumoricon/site/attendee/search/PrintBadgePresenter.java
@@ -26,6 +26,7 @@ public class PrintBadgePresenter extends BadgePrintingPresenter {
         Attendee attendee = attendeeRepository.findOne(id);
         if (attendee != null) {
             log.info("{} displayed Attendee {}", view.getCurrentUsername(), attendee);
+            printBadges(view, Arrays.asList(attendee));
             view.showAttendee(attendee);
         } else {
             log.error("{} tried to display Attendee id {} and it was not found", view.getCurrentUsername(), id);